feat: add jetson overlay
Add NVIDIA Jetson family overlay.
