+++ .github/renovate.json
@@ -32,13 +32,6 @@
     ],
     "packageRules": [
         {
-            "matchPackageNames": [
-                "u-boot/u-boot",
-                "ARM-software/arm-trusted-firmware"
-            ],
-            "versioning": "regex:^(?<major>\\d+)\\.(?<minor>\\d+)\\.?(?<patch>\\d+)?$"
-        },
-        {
             "matchPackagePatterns": [
                 "*"
             ],
