+++ artifacts/u-boot/pkg.yaml
@@ -10,10 +10,10 @@ dependencies:
   - stage: base
 steps:
   - sources:
-      - url: https://ftp.denx.de/pub/u-boot/u-boot-{{ .uboot_version }}.tar.bz2
+      - url: https://github.com/OE4T/u-boot-tegra/archive/refs/tags/v2022.07.tar.gz
         destination: u-boot.tar.bz2
-        sha256: "{{ .uboot_sha256 }}"
-        sha512: "{{ .uboot_sha512 }}"
+        sha256: e8f03d0b3dc2d70db5833f0e6f3ac305c0de0d71ff823c7d6210edee68e29a46
+        sha512: 986e79cfe56878f17750144c8e21f03a9f802d94b7e39ca7268336422f8ab99268d85a4debe5c8f48e25a31f8e1e37e02b3fad448c2737dee286fb07486a06ba
     env:
       SOURCE_DATE_EPOCH: {{ .BUILD_ARG_SOURCE_DATE_EPOCH }}
     prepare:
@@ -23,15 +23,15 @@ steps:
 
         tar xf u-boot.tar.bz2 --strip-components=1
       - |
-        make <board>_defconfig
+        make p3450-0000_defconfig
         sed -i "s/CONFIG_TOOLS_LIBCRYPTO=y/# CONFIG_TOOLS_LIBCRYPTO is not set/" .config
     build:
       - |
         make -j $(nproc) HOSTLDLIBS_mkimage="-lssl -lcrypto"
     install:
       - |
-        mkdir -p /rootfs/artifacts/arm64/u-boot/board
-        cp -v u-boot.bin /rootfs/artifacts/arm64/u-boot/board
+        mkdir -p /rootfs/artifacts/arm64/u-boot/jetson_nano
+        cp -v u-boot.bin /rootfs/artifacts/arm64/u-boot/jetson_nano
 finalize:
   - from: /rootfs
     to: /rootfs
