+++ installers/jetson_nano/src/go.mod
@@ -1,10 +1,11 @@
-module board
+module jetson_name
 
 go 1.22.1
 
 require (
 	github.com/siderolabs/go-copy v0.1.0
 	github.com/siderolabs/talos/pkg/machinery v1.7.0-alpha.1
+	golang.org/x/sys v0.18.0
 )
 
 require gopkg.in/yaml.v3 v3.0.1 // indirect
