+++ profiles/jetson_nano/jetson_nano.yaml
@@ -3,7 +3,7 @@ platform: metal
 secureboot: false
 output:
   kind: image
-  outFormat: raw
+  outFormat: .xz
   imageOptions:
     diskSize: 1306525696
     diskFormat: raw
